src/requirements.txt
python-dotenv==1.0.1
langchain==0.1.20
langchain-openai==0.1.7
langchain-core==0.1.52
nemoguardrails==0.9.0
fastapi==0.111.0
uvicorn==0.30.1
nest-asyncio==1.6.0
openai==1.34.0
pydantic==2.8.2
pydantic-settings==2.4.0
pandas==2.2.2
openpyxl==3.1.5